+++ packages/core/src/mock/validations.js
@@ -245,7 +245,7 @@ function customValidationSingleMessage(errors) {
     .join(". ");
 }
 
-function validationSingleMessage(schema, data, errors) {
+function validationSingleMessage(schema, data = {}, errors) {
   const formattedJson = betterAjvErrors(schema, data, errors, {
     format: "js",
   });