+++ packages/plugin-openapi/test/fixtures/api-users-collection-no-variants/openapi/api.js
@@ -1,0 +1,29 @@
+const deepMerge = require("deepmerge");
+
+const openApiDocument = require("../../../openapi/users");
+
+module.exports = [
+  {
+    basePath: "/api",
+    collection: {
+      id: "users",
+    },
+    document: deepMerge(openApiDocument, {
+      paths: {
+        "/users": {
+          get: {
+            responses: {
+              "200": {
+                content: {
+                  "application/json": {
+                    examples: undefined,
+                  },
+                },
+              },
+            },
+          },
+        },
+      },
+    }),
+  },
+];